+++ lib/generators/magick/install/install_generator.rb
@@ -2,7 +2,7 @@
 
 module Magick
   module Generators
-    class InstallGenerator < Rails::Generators::Base
+    class InstallGenerator < ::Rails::Generators::Base
       source_root File.expand_path('templates', __dir__)
 
       desc 'Creates a Magick configuration file at config/initializers/magick.rb'